fel: exit with error if no FEL device at busnum/devnum
The fel code confusingly parses the bus and dev numbers output by lsusb
as octal (in stroul()) since they have leading zeros.  This causes fel
to select the first available VID/PID matching device and causes
confusing failures when using fel on multiple devices simultaneously.

A better approach is to assume the user meant to connect only to the
explicit busnum and devnum when provided and fail if that is not a fel
device.

Signed-off-by: Jason Abele <jason@nextthing.co>